pyproject.toml
[project]
name = "nl2sql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.15",
    "ipykernel>=6.29.5",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.7",
    "langchain-openai>=0.3.27",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
]
